requirements.txt
matplotlib==3.10.8
numpy==2.3.5
pandas==2.3.3
Pillow==12.0.0
scikit_learn==1.4.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
